requirements.txt
streamlit==1.55.0
pandas==2.3.3
plotly==6.6.0
duckdb==1.5.0
requests==2.32.5
matplotlib==3.10.8
numpy==2.4.3
python-dotenv==1.2.2
